LGA-2350 Moved make_dir method to common steps. Added sleep time for check_accessibility to prevent Javascript exceptions. New error logging when accessbility issues are found.
